requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blis==1.0.1
breadability==0.1.20
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
distro==1.9.0
docopt==0.6.2
editdistpy==0.1.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
gensim==4.3.3
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hyperframe==6.0.1
idna==3.10
inexactsearch==1.0.2
jellyfish==1.1.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
langdetect==1.0.9
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
Metaphone==0.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.52.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
portalocker==2.10.1
preshed==3.0.9
propcache==0.2.0
protobuf==5.28.2
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.0
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
segtok==1.5.11
sentence-transformers==3.2.1
shellingham==1.5.4
silpa-common==0.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soundex==1.1.3
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sumy==0.11.0
sympy==1.13.1
symspellpy==6.7.8
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wordcloud==1.9.3
wrapt==1.16.0
yake==0.4.8
yarl==1.15.5
